Formatting changes

+++ src/engine/math_util.c
@@ -45,20 +45,20 @@ Vec3i gVec3iZero = {     0,     0,     0 };
 Vec3s gVec3sOne  = {     1,     1,     1 };
 
 /// From Wiseguy
-static inline int asm_roundf(float in) {
-    float tmp;
-    int out;
+static inline s32 asm_roundf(f32 in) {
+    f32 tmp;
+    s32 out;
     __asm__("round.w.s %0,%1" : "=f" (tmp) : "f" (in));
     __asm__("mfc1 %0,%1" : "=r" (out) : "f" (tmp));
     return out;
 }
 
-inline f32 roundf(f32 x) {
+inline s32 roundf(f32 x) {
     return asm_roundf(x);
 }
 
-static inline float asm_absf(float in) {
-    float out;
+static inline f32 asm_absf(f32 in) {
+    f32 out;
     __asm__("abs.s %0,%1" : "=f" (out) : "f" (in));
     return out;
 }
